Changed some things to the objectpool
