+++ Assets/Scripts/ObjectPool/Demo/GameObject (1).prefab
@@ -16,7 +16,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &8429155717268920460
 Transform:
   m_ObjectHideFlags: 0
