--- Assets/Scripts/ObjectPool/IPoolableObject.cs
@@ -1,8 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface IPoolableObject
-{
-    public void ResetValues();
-}
