--- Assets/Scripts/ObjectPool/IPoolableObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1aecc63f93ee87b4494344ee24d7d862
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
