--- Assets/Scripts/ObjectPool/ObjectPool1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8352a3e89d693c848b5485b5702c26e9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
