--- Assets/Scripts/PAXSpawner/PassengerSpawnManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 292215c239bd9af4c91bf895918f0abf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
