--- Assets/Scripts/PAXSpawner/PassengerSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c05f5f78a76bd834681b760814e879b7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
